+++ lib/belt.rb
@@ -2,7 +2,11 @@
 
 require_relative "belt/version"
 require_relative "belt/parameters"
-require_relative "belt/application_controller"
 
 module Belt
+  class AuthenticationError < StandardError; end
+  class RecordNotFound < StandardError; end
+  class ActionNotFound < StandardError; end
 end
+
+require_relative "belt_controller/base"