+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.0.2 (2026-06-10)
 
-- Add ApplicationController with callbacks, strong parameters, error handling, and CORS
+- Add BeltController::Base with callbacks, strong parameters, error handling, and CORS
 - Add activeitem, lambda_loadout, and s3arch as dependencies
 - Fix conflicting VERSION constants